kernel: give interrupt source, controller and instance their own widths
An interrupt source number crossed five layers and each had picked its own
width, so widening any single one changed nothing. icid_t made this worse by
typing three different things at once: the controller id, the controller
instance, and the IRQ number itself.

Widening the typedef could not have helped in any case, because the identity
was not stored in it. The controller id lived in ic_Node.ln_Type and the
per-IRQ pair in the handler chain's own lh_Type and l_pad - all UBYTE by the
exec ABI, and so capped at 255 no matter how wide the type became. A machine
with more controller instances than that aliased one onto another and pointed
an IRQ at the wrong hardware.

The three spaces are now separate types - irqid_t, icid_t, icinstid_t, with
icintrid_t the packed pair - derived in kernel_irqtypes.h from bit widths an
arch declares in its <aros/cpu.h>, alongside HW_IRQ_BASE. The identity moves
into IntrController.ic_Id and KernelInt, out of the exec fields. ICINTR_MAKE()
joins the accessors so the split is no longer open-coded at a fixed 8 bits.

The defaults reproduce the widths these APIs have always had, so m68k and the
other narrow ports are untouched; x86_64 takes 16/16 for controller id and
instance. The public LVOs already passed uint32_t and ULONG and are unchanged.

riscv64-opensbi raises HW_IRQ_COUNT to the PLIC maximum of 1024 and drops the
uint8_t its ictl_* calls truncated the source to, which cost it every number
above 255 with no diagnostic; pcidt can now ask for all 32 message vectors a
DesignWare bridge has rather than 8.

+++ arch/riscv64-opensbi/kernel/modifyirq.c
@@ -114,7 +114,7 @@ struct krnMSIController *krnFindMSIController(unsigned int irq)
              * so there is nothing else to unmask it, and until it is
              * the controller can receive messages that never arrive.
              */
-            ictl_enable_irq((uint8_t)irq, KernelBase);
+            ictl_enable_irq(irq, KernelBase);
 
             D(bug("[KRN] %s: source %u collects %u vector(s) from %u\n",
                   __func__, irq, __msi_ctrl[i].count, __msi_ctrl[i].base);)
